minor fixes added
