+++ db/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 
 [dependencies]
 rocksdb = { version = "0.21", default-features = false, features = ["lz4", "multi-threaded-cf"] }
+actix-web = "4"
